Organize files ::  Remove unnecessary files
